+++ uitools/toolkitcpp/src/Esri/ArcGISRuntime/Toolkit/register.cpp
@@ -76,14 +76,14 @@
 
 /*!
   \headerfile Esri/ArcGISRuntime/Toolkit/register
-  \inmodule ArcGISRuntimeToolkit
+    \inmodule ArcGISRuntimeToolkit
 
-  This file contains the registration function required to register the toolkit
-  with the `QQmlEngine`.
+    This file contains the registration function required to register the toolkit
+    with the `QQmlEngine`.
 
-  Please refer to
-  \c README.md for more information on workflows.
- */
+    Please refer to
+    \c README.md for more information on workflows.
+   */
 
 /*!
   \fn void Esri::ArcGISRuntime::Toolkit::registerComponents(QQmlEngine& engine)
@@ -101,8 +101,8 @@ namespace Esri::ArcGISRuntime::Toolkit {
 
     constexpr char const* NAMESPACE = "Esri.ArcGISRuntime.Toolkit.Controller";
 
-    constexpr int VERSION_MAJOR = 200;
-    constexpr int VERSION_MINOR = 2;
+    constexpr int VERSION_MAJOR = 3;
+    constexpr int VERSION_MINOR = 0;
 
     QPointer<AuthenticatorController> s_authenticatorController;
 
@@ -114,9 +114,9 @@ namespace Esri::ArcGISRuntime::Toolkit {
 
       Provided is the overloaded method:
 
-      \code
+  \code
       void registerComponentImpl(<CreationType> creationType, int majorVersion, int minorVersion, const char* name)
-      \endcode
+  \endcode
 
       And the three currently accepted values for the \a creationType parameter.
 
@@ -173,7 +173,7 @@ namespace Esri::ArcGISRuntime::Toolkit {
     /*
      \internal
      \brief Function for registration. Registers the C++ type Foo as
-     Foo in QML with the 100.10, 200.0 version and namespace information.
+     Foo in QML with the 3.0 version and namespace information.
 
   \list
       \li \a Determines how the type is instantiated in QML. Choose between CreationType::Creatable, CreationType::Uncreatable and CreationType::Interface.
@@ -181,9 +181,9 @@ namespace Esri::ArcGISRuntime::Toolkit {
   \endlist
 
       Example call:
-      \code
+  \code
       registerComponent<LocatorSearchSource>(CreationType::Uncreatable);
-      \endcode
+  \endcode
  */
     template <typename T, typename CType = CreationType::Creatable_>
     void registerComponent(CType creationType = CType{})
@@ -191,32 +191,7 @@ namespace Esri::ArcGISRuntime::Toolkit {
       static_assert(std::is_base_of<QObject, T>::value, "Must inherit QObject");
       auto name = QString{T::staticMetaObject.className()};
       name.remove("Esri::ArcGISRuntime::Toolkit::");
-      // register component on version 100
-      CreationType::registerComponentImpl<T>(creationType, 100, 10, name.toLatin1());
-      // register component on version 200
-      CreationType::registerComponentImpl<T>(creationType, 200, 0, name.toLatin1());
-    }
-
-    /*
-     \internal
-     \brief Ensures a Module revision is available from 100.10 and 200.0 onwards
-     to the current version of the Toolkit.
- */
-    void registerModuleRevisions()
-    {
-      constexpr int MAJOR_VERSION_100 = 100;
-      constexpr int START_MINOR_VERSION_100 = 10;
-      constexpr int END_MINOR_VERSION_100 = 15;
-      // register version 100
-      for (int i = START_MINOR_VERSION_100; i <= END_MINOR_VERSION_100; ++i)
-      {
-        qmlRegisterModule(NAMESPACE, MAJOR_VERSION_100, i);
-      }
-      // register version 200 onwards
-      for (int i = 0; i <= VERSION_MINOR; ++i)
-      {
-        qmlRegisterModule(NAMESPACE, VERSION_MAJOR, i);
-      }
+      CreationType::registerComponentImpl<T>(creationType, VERSION_MAJOR, VERSION_MINOR, name.toLatin1());
     }
 
   } // namespace
@@ -226,7 +201,7 @@ namespace Esri::ArcGISRuntime::Toolkit {
     appEngine.addImageProvider(BasemapGalleryImageProvider::PROVIDER_ID, BasemapGalleryImageProvider::instance());
     appEngine.addImageProvider(PopupAttachmentImageProvider::PROVIDER_ID, PopupAttachmentImageProvider::instance());
     appEngine.addImportPath(ESRI_COM_PATH);
-    registerModuleRevisions();
+    qmlRegisterModule(NAMESPACE, VERSION_MAJOR, VERSION_MINOR);
     registerComponent<AuthenticatorController>(CreationType::Singleton);
     registerComponent<AttachmentsPopupElementViewController>();
     registerComponent<BarChartPopupMediaItem>();
@@ -268,8 +243,7 @@ namespace Esri::ArcGISRuntime::Toolkit {
 
     // Register ArcGISRuntime types with toolkit.
     qRegisterMetaType<Point>("Esri::ArcGISRuntime::Point");
-    qmlRegisterAnonymousType<MapQuickView>(NAMESPACE, 100);
-    qmlRegisterAnonymousType<MapQuickView>(NAMESPACE, 200);
+    qmlRegisterAnonymousType<MapQuickView>(NAMESPACE, VERSION_MAJOR);
   }
 
 } // Esri::ArcGISRuntime::Toolkit